+++ packages/core/src/public-utils/reorder.ts
@@ -1,6 +1,7 @@
 /**
- * Reorder a provided `list`
- * Returns a new array and does not modify the original array
+ * A function that will reorder an array (`list`).
+ * `reorder` returns a new array with reordered items and does not
+ *  modify the original array. The items in the array are also not modified.
  */
 export function reorder<Value>({
 	list,
@@ -12,7 +13,10 @@ export function reorder<Value>({
 	finishIndex: number;
 }): Value[] {
 	if (startIndex === -1 || finishIndex === -1) {
-		return list;
+		// Making this function consistently return a new array reference.
+		// This is consistent with .toSorted() which always returns a new array
+		// even when it does not do anything
+		return Array.from(list);
 	}
 
 	const result = Array.from(list);