+++ packages/react-drop-indicator/examples/box-playground.tsx
@@ -5,7 +5,7 @@
 
 import { css, jsx } from '@compiled/react';
 
-import { Stack } from '@atlaskit/primitives';
+import { Stack } from '@atlaskit/primitives/compiled';
 import { token } from '@atlaskit/tokens';
 
 import { AppearanceExample } from './constellation/box/box-appearance';